#154 Alterando fluxo da requisição, que agora é feito na propriedade
