+++ plantio/urls.py
@@ -13,11 +13,5 @@ urlpatterns = [
                 r'agrotoxico/$', aplicacao.AplicacaoAgrotoxicoAPIView.as_view(),
                 name='plantio-associar'),
         ])),
-
-        url(r'^(?P<idPropriedade>[0-9A-Za-z]{8}-[0-9A-Za-z]{4}-4[0-9A-Za-z]{3}-[89ABab][0-9A-Za-z]{3}-[0-9A-Za-z]{12})/historico/', include([
-            url(
-                r'^propriedade/$', plantio.PlantioHistoricoPropriedadeAPIView.as_view(),
-                name='plantio-historico-propriedade'),
-        ])),
     ]))
 ]
