+++ plantio/views/plantio.py
@@ -1,9 +1,7 @@
 from plantio.models import Plantio
 from plantio.serializers.plantio import PlantioSerializer
 
-from rest_framework.generics import CreateAPIView, ListAPIView
-
-from talhao.models import Talhao
+from rest_framework.generics import CreateAPIView
 
 
 class PlantioAPIView(CreateAPIView):
@@ -11,12 +9,3 @@ class PlantioAPIView(CreateAPIView):
 
     def get_queryset(self):
         return Plantio.objects.all()
-
-
-class PlantioHistoricoPropriedadeAPIView(ListAPIView):
-    serializer_class = PlantioSerializer
-    lookup_field = 'idPropriedade'
-
-    def get_queryset(self):
-        return Plantio.objects.filter(
-            talhao__in=Talhao.objects.filter(**self.kwargs).values_list('idTalhao'))
