+++ propriedade/tests/views/test_propriedade.py
@@ -6,11 +6,17 @@ from rest_framework.reverse import reverse_lazy
 
 from parameterized import parameterized
 
+from plantio.models import Plantio
+from plantio.tests.recipes import plantio
+
 from propriedade.models import Propriedade
+from propriedade.tests.recipes import propriedade
 
 from produtor.tests.recipes import produtor
 
 from tecnico.tests.recipes import tecnico
+
+from talhao.tests.recipes import talhao
 
 
 class PropriedadeAPIViewTest(APITestMixin, TestCase):
@@ -111,3 +117,33 @@ class PropriedadeAPIViewTest(APITestMixin, TestCase):
             'CEP deve possuir 8 digítos numéricos.',
             response.json()['cep']
         )
+
+
+class PropriedadeHistoricoPlantioAPIView(APITestMixin, TestCase):
+
+    def setUp(self):
+        self.propriedade = propriedade.make()
+        self.url = reverse_lazy(
+            "propriedade-historico-plantio", kwargs={'idPropriedade': self.propriedade.idPropriedade})
+        self.talhoes = talhao.make(idPropriedade=self.propriedade, _quantity=3)
+        talhao.make(idPropriedade=self.propriedade, _quantity=2)
+        for t in self.talhoes:
+            plantio.make(talhao=t)
+
+    def test_lista_plantios_do_talhao(self):
+        response = self.client.get(self.url)
+        self.assertEqual(response.status_code, 200, response.json())
+        self.assertEqual(len(response.json()), 3)
+
+    def test_nao_lista_plantio_de_outra_propriedade(self):
+        url = reverse_lazy(
+            "propriedade-historico-plantio", kwargs={'idPropriedade': self.propriedade.idPropriedade})
+        propriedade_diferente = propriedade.make()
+        talhoes = talhao.make(idPropriedade=propriedade_diferente, _quantity=3)
+        for t in talhoes:
+            plantio.make(talhao=t)
+
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200, response.json())
+        self.assertEqual(Plantio.objects.count(), 6)
+        self.assertEqual(len(response.json()), 3)
