+++ propriedade/views/propriedade.py
@@ -1,7 +1,12 @@
+from plantio.models import Plantio
+from plantio.serializers.plantio import PlantioSerializer
+
 from propriedade.models import Propriedade
 from propriedade.serializers.propriedade import PropriedadeSerializer
 
-from rest_framework.generics import CreateAPIView
+from rest_framework.generics import CreateAPIView, ListAPIView
+
+from talhao.models import Talhao
 
 
 class PropriedadeAPIView(CreateAPIView):
@@ -9,3 +14,11 @@ class PropriedadeAPIView(CreateAPIView):
 
     def get_queryset(self):
         return Propriedade.objects.all()
+
+class PropriedadeHistoricoPlantioAPIView(ListAPIView):
+    serializer_class = PlantioSerializer
+    lookup_field = 'idPropriedade'
+
+    def get_queryset(self):
+        return Plantio.objects.filter(
+            talhao__in=Talhao.objects.filter(**self.kwargs).values_list('idTalhao'))
